FY-96/feat: add GameHistory class
